Switch to i64 everywhere

+++ src/p01.rs
@@ -1,6 +1,6 @@
 use crate::utils;
 
-fn solve(v: &[i32]) -> i32 {
+fn solve(v: &[i64]) -> i64 {
     let mut count = 0;
     for i in 1..v.len() {
         if v[i] > v[i - 1] {
@@ -10,12 +10,12 @@ fn solve(v: &[i32]) -> i32 {
     count
 }
 
-pub fn solve_1() -> i32 {
+pub fn solve_1() -> i64 {
     let v = utils::read_ints("data/01.in", "\n").unwrap();
     solve(&v)
 }
 
-pub fn solve_2() -> i32 {
+pub fn solve_2() -> i64 {
     let v = utils::read_ints("data/01.in", "\n").unwrap();
     let mut w = Vec::new();
     for i in 0..v.len() {
