+++ src/p06.rs
@@ -2,7 +2,7 @@ use std::error::Error;
 
 use crate::utils;
 
-pub fn solve(days: i32) -> Result<i64, Box<dyn Error>> {
+pub fn solve(days: i64) -> Result<i64, Box<dyn Error>> {
     let counters = utils::read_ints("data/06.in", ",")?;
 
     let mut frequency: Vec<i64> = vec![0; 9];
